storage tests

+++ js/storage.js
@@ -5,9 +5,11 @@ export const setItem = (name, item) => {
 };
 
 export const getItem = (name) => {
+    if (!window.localStorage) return false;
     return window.localStorage.getItem(name);
 };
 
 export const clearStorage = () => {
+    if (!window.localStorage) return false;
     window.localStorage.clear();
 };
